Add setting texture during runtime
